back-general-feat: mysql configuration and railway deployment

+++ backend/src/main/java/org/petzonalize/backend/repository/CustomizableRepository.java
@@ -3,6 +3,6 @@ package org.petzonalize.backend.repository;
 import org.petzonalize.backend.entity.Customizable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomizableRepository extends JpaRepository<Customizable, Long> {
+public interface CustomizableRepository extends JpaRepository<Customizable, Integer> {
 
 }
