+++ backend/src/main/java/org/petzonalize/backend/repository/ProductSizeRepository.java
@@ -3,6 +3,6 @@ package org.petzonalize.backend.repository;
 import org.petzonalize.backend.entity.ProductSize;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductSizeRepository extends JpaRepository<ProductSize, Long> {
+public interface ProductSizeRepository extends JpaRepository<ProductSize, Integer> {
 
 }
